# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Analytics Model API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /odds/capture:
    post:
      operationId: captureOdds
      tags: [dashboard]
      summary: Make one explicit Odds API capture
      description: Performs one server-side request only. It never polls or retries and never exposes the API key.
      responses:
        "200":
          description: Capture completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OddsCaptureResult"
        "502":
          description: Capture failed
  /odds/audit:
    get:
      operationId: listOddsAudits
      tags: [dashboard]
      summary: Read persisted per-event Odds API matching audits
      description: Returns parsed provider fields and matching diagnostics without raw payloads or keyed URLs.
      parameters:
        - name: requestId
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: outcome
          in: query
          required: false
          schema:
            type: string
            enum: [matched_saved, matched_post_kickoff_skipped, unmatched]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        "200":
          description: Persisted per-event audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OddsEventAudit"
  /data-sync/schedule:
    post:
      operationId: syncSchedule
      tags: [nfl]
      summary: Persist the current and next two exposed NFL weeks
      description: Performs an explicit ESPN schedule sync and only upserts games; it never deletes schedule history or calls the Odds API.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleSyncRequest"
      responses:
        "200":
          description: Schedule synchronization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSyncResult"
        "502":
          description: Schedule synchronization failed
...
  /consumer/dashboard:
    get:
      operationId: getConsumerDashboard
      tags: [nfl]
      summary: Read the persisted consumer game dashboard
      description: Returns consumer-safe, persisted data only. It never starts data synchronization or model computation.
      responses:
        "200":
          description: Consumer dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerDashboard"
        "503":
          $ref: "#/components/responses/ConsumerDataUnavailable"
  /consumer/games:
    get:
      operationId: listConsumerGames
      tags: [nfl]
      summary: Read consumer-safe persisted games
      parameters:
        - name: season
          in: query
          schema: { type: integer, minimum: 2020 }
        - name: week
          in: query
          schema: { type: integer, minimum: 1, maximum: 22 }
      responses:
        "200":
          description: Consumer games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerGames"
        "400":
          $ref: "#/components/responses/ConsumerInvalidRequest"
        "503":
          $ref: "#/components/responses/ConsumerDataUnavailable"
  /consumer/games/{gameId}:
    get:
      operationId: getConsumerGame
      tags: [nfl]
      summary: Read a consumer-safe persisted game detail
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Consumer game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerGameDetail"
        "404":
          $ref: "#/components/responses/ConsumerGameNotFound"
        "503":
          $ref: "#/components/responses/ConsumerDataUnavailable"
  /consumer/performance:
    get:
      operationId: getConsumerPerformance
      tags: [nfl]
      summary: Read consumer-safe persisted performance
      responses:
        "200":
          description: Consumer performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerPerformance"
        "503":
          $ref: "#/components/responses/ConsumerDataUnavailable"
  /consumer/trends:
    get:
      operationId: getConsumerTrends
      tags: [nfl]
      summary: Read consumer-safe performance trends
      responses:
        "200":
          description: Consumer trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerTrends"
        "503":
          $ref: "#/components/responses/ConsumerDataUnavailable"
...
components:
  responses:
    ConsumerDataUnavailable:
      description: Consumer data is temporarily unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ConsumerError" }
    ConsumerInvalidRequest:
      description: Invalid consumer request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ConsumerError" }
    ConsumerGameNotFound:
      description: Consumer game unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ConsumerError" }
  schemas:
    ConsumerError:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string, enum: [consumer_data_unavailable, invalid_request, game_not_found] }
    ConsumerTeam:
      type: object
      required: [name, abbreviation, logoUrl]
      properties:
        name: { type: string }
        abbreviation: { type: string }
        logoUrl: { type: ["string", "null"] }
    ConsumerMarketQuote:
      type: object
      required: [sportsbook, point, price, capturedAt]
      properties:
        sportsbook: { type: string }
        point: { type: ["number", "null"] }
        price: { type: number }
        capturedAt: { type: ["string", "null"], format: date-time }
    ConsumerGame:
      type: object
      required: [gameId, season, week, kickoffTime, gameStatus, venue, matchup, finalScore, prediction, market, dataConfidence, availability]
      properties:
        gameId: { type: string }
        season: { type: integer }
        week: { type: integer }
        kickoffTime: { type: ["string", "null"], format: date-time }
        gameStatus: { type: string }
        venue: { type: ["string", "null"] }
        matchup:
          type: object
          required: [home, away]
          properties:
            home: { $ref: "#/components/schemas/ConsumerTeam" }
            away: { $ref: "#/components/schemas/ConsumerTeam" }
        finalScore:
          type: ["object", "null"]
          properties:
            home: { type: integer }
            away: { type: integer }
        prediction:
          type: ["object", "null"]
          properties:
            modelLabel: { type: string, const: Gridline Production Model }
            projectedHomeScore: { type: ["number", "null"] }
            projectedAwayScore: { type: ["number", "null"] }
            projectedMargin: { type: ["number", "null"] }
            projectedTotal: { type: ["number", "null"] }
            homeWinProbability: { type: ["number", "null"] }
            awayWinProbability: { type: ["number", "null"] }
        market:
          type: object
          required: [spread, moneyline, total]
          properties:
            spread: { anyOf: [{ $ref: "#/components/schemas/ConsumerMarketQuote" }, { type: "null" }] }
            moneyline: { anyOf: [{ $ref: "#/components/schemas/ConsumerMarketQuote" }, { type: "null" }] }
            total: { anyOf: [{ $ref: "#/components/schemas/ConsumerMarketQuote" }, { type: "null" }] }
        dataConfidence:
          type: object
          required: [label, score, reason]
          properties:
            label: { type: string, enum: [Updating, Limited, Moderate, Standard] }
            score: { type: ["number", "null"] }
            reason: { type: ["string", "null"] }
        availability:
          type: object
          required: [prediction, market]
          properties:
            prediction: { type: ["string", "null"] }
            market: { type: ["string", "null"] }
    ConsumerGames:
      type: object
      required: [status, games]
      properties:
        status: { type: string, enum: [available, unavailable] }
        games:
          type: array
          items: { $ref: "#/components/schemas/ConsumerGame" }
    ConsumerDashboard:
      type: object
      required: [status, games, note]
      properties:
        status: { type: string, enum: [available, unavailable] }
        games:
          type: array
          items: { $ref: "#/components/schemas/ConsumerGame" }
        note: { type: string }
    ConsumerGameDetail:
      allOf:
        - $ref: "#/components/schemas/ConsumerGame"
        - type: object
          required: [weather, analysis]
          properties:
            weather:
              type: ["object", "null"]
              additionalProperties: true
            analysis:
              type: object
              required: [drivers, availability]
              properties:
                drivers: { type: array, items: { type: string } }
                availability:
                  type: object
                  required: [weather, personnel]
                  properties:
                    weather: { type: ["string", "null"] }
                    personnel: { type: ["string", "null"] }
    ConsumerPerformance:
      type: object
      required: [status, officialPredictions, gradedPredictions, byFamily, breakdowns, note]
      properties:
        status: { type: string, enum: [available, unavailable] }
        officialPredictions: { type: integer }
        gradedPredictions: { type: integer }
        byFamily: { type: object, additionalProperties: true }
        breakdowns: { type: object, additionalProperties: true }
        note: { type: string }
    ConsumerTrends:
      type: object
      required: [status, byWeek, byConfidence, byEdge, note]
      properties:
        status: { type: string, enum: [available, unavailable] }
        byWeek: { type: array, items: { type: object, additionalProperties: true } }
        byConfidence: { type: array, items: { type: object, additionalProperties: true } }
        byEdge: { type: array, items: { type: object, additionalProperties: true } }
        note: { type: string }
...
    OddsCaptureResult:
      type: object
      required:
        [status, requestedAt, requestId, requestCount, recordsReceived, snapshotsCreated,
         auditedEvents, duplicateSnapshots, unmatchedEvents, skippedPostKickoff,
         missingMarkets, failedSportsbooks, creditsUsed, creditsRemaining, error]
      properties:
        status: { type: string, enum: [success, failed, not_configured, skipped] }
        requestedAt: { type: string, format: date-time }
        requestId: { type: ["integer", "null"] }
        requestCount: { type: integer }
        recordsReceived: { type: integer }
        auditedEvents: { type: integer }
        snapshotsCreated: { type: integer }
        duplicateSnapshots: { type: integer }
        unmatchedEvents: { type: integer }
        skippedPostKickoff: { type: integer }
        missingMarkets:
          type: array
          items: { type: string }
        failedSportsbooks:
          type: array
          items: { type: string }
        creditsUsed: { type: ["integer", "null"] }
        creditsRemaining: { type: ["integer", "null"] }
        skipReason: { type: ["string", "null"] }
        error: { type: ["string", "null"] }
    OddsAuditCandidate:
      type: object
      required: [gridlineGameId, kickoffTime, timeDifferenceMinutes]
      properties:
        gridlineGameId: { type: string }
        kickoffTime: { type: ["string", "null"], format: date-time }
        timeDifferenceMinutes: { type: ["number", "null"] }
    OddsEventAudit:
      type: object
      required:
        [id, requestId, eventIndex, providerEventId, providerHomeTeam,
         providerAwayTeam, providerKickoffTime, normalizedHomeTeam,
         normalizedAwayTeam, candidateGridlineGames, matchedGridlineGameId,
         matchedGridlineKickoff, outcome, reason, observationsReceived,
         observationsSaved, duplicateObservations, rejectedObservations, auditedAt]
      properties:
        id: { type: integer }
        requestId: { type: integer }
        eventIndex: { type: integer }
        providerEventId: { type: ["string", "null"] }
        providerHomeTeam: { type: ["string", "null"] }
        providerAwayTeam: { type: ["string", "null"] }
        providerKickoffTime: { type: ["string", "null"], format: date-time }
        normalizedHomeTeam: { type: ["string", "null"] }
        normalizedAwayTeam: { type: ["string", "null"] }
        candidateGridlineGames:
          type: array
          items: { $ref: "#/components/schemas/OddsAuditCandidate" }
        matchedGridlineGameId: { type: ["string", "null"] }
        matchedGridlineKickoff: { type: ["string", "null"], format: date-time }
        outcome:
          type: string
          enum: [matched_saved, matched_post_kickoff_skipped, unmatched]
        reason:
          type: string
          enum:
            [saved_observation, duplicate_observation, no_observations,
             postkickoff, invalid_fields, no_matching_teams, missing_schedule,
             outside_tolerance, ambiguity, other]
        observationsReceived: { type: integer }
        observationsSaved: { type: integer }
        duplicateObservations: { type: integer }
        rejectedObservations: { type: integer }
        auditedAt: { type: string, format: date-time }
    ScheduleSyncRequest:
      type: object
      properties:
        season: { type: integer, minimum: 2020 }
        currentWeek: { type: integer, minimum: 1, maximum: 22 }
    ScheduleSyncResult:
      type: object
      required: [status, season, currentWeek, weeks, gamesByWeek, totalGames, failures]
      properties:
        status: { type: string, enum: [success, partial, failed] }
        season: { type: integer }
        currentWeek: { type: integer }
        weeks:
          type: array
          items: { type: integer }
        gamesByWeek:
          type: object
          additionalProperties: { type: integer }
        totalGames: { type: integer }
        failures:
          type: array
          items: { type: string }
...