# lib/api-zod/src/generated/types/consumerError.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NFL Analytics Model API
 * OpenAPI spec version: 0.1.0
 */
import type { ConsumerErrorCode } from './consumerErrorCode';

export interface ConsumerError {
  error: string;
  code: ConsumerErrorCode;
}