# lib/api-zod/src/generated/types/consumerGame.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NFL Analytics Model API
 * OpenAPI spec version: 0.1.0
 */
import type { ConsumerGameAvailability } from './consumerGameAvailability';
import type { ConsumerGameDataConfidence } from './consumerGameDataConfidence';
import type { ConsumerGameFinalScore } from './consumerGameFinalScore';
import type { ConsumerGameMarket } from './consumerGameMarket';
import type { ConsumerGameMatchup } from './consumerGameMatchup';
import type { ConsumerGamePrediction } from './consumerGamePrediction';

export interface ConsumerGame {
  gameId: string;
  season: number;
  week: number;
  /** @nullable */
  kickoffTime: Date | null;
  gameStatus: string;
  /** @nullable */
  venue: string | null;
  matchup: ConsumerGameMatchup;
  /** @nullable */
  finalScore: ConsumerGameFinalScore;
  /** @nullable */
  prediction: ConsumerGamePrediction;
  market: ConsumerGameMarket;
  dataConfidence: ConsumerGameDataConfidence;
  availability: ConsumerGameAvailability;
}