# lib/api-zod/src/generated/types/consumerTeam.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NFL Analytics Model API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsumerTeam {
  name: string;
  abbreviation: string;
  /** @nullable */
  logoUrl: string | null;
}